requirements.txt
langchain
sentence_transformers
torch
sentencepiece
sentence_transformers
accelerate
chromadb
pypdf
tiktoken
streamlit
fastapi
uvicorn
python-multipart
aiofiles